1) Добавил beta версию подкачки комнат с локального/remote хранилища

+++ app/src/main/java/com/example/quick_cheque/data/local/entity/RoomEntity.kt
@@ -5,7 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "room")
 data class RoomEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int? = null,
+    @PrimaryKey(autoGenerate = true)
+    val id: Long? = null,
     val titleRoom: String,
-    val ownerId: Int,
+    val ownerId: Long,
 )
